requirements.txt
python-dotenv===1.0.1
mysql-connector-python===8.4.0
psycopg2-binary===2.9.9
cx-oracle===8.3.0
pytest==7.4.4
blue===0.9.1
